move main.go bitfan to cmd/bitfan
