+++ .gitignore
@@ -25,7 +25,7 @@ _testmain.go
 
 tester
 examples.d/twitter.conf
-bitfan
+
 releases
 .DS_Store
