+++ cmd/bitfan/commands/cmd_test.go
@@ -1,4 +1,4 @@
-package cli
+package commands
 
 import "testing"
