+++ cmd/bitfan/commands/conf.go
@@ -1,4 +1,4 @@
-package cli
+package commands
 
 import (
 	"fmt"
