+++ cmd/bitfan/commands/plugins.go
@@ -1,4 +1,4 @@
-package cli
+package commands
 
 import (
 	"github.com/vjeantet/bitfan/core"
