+++ cmd/bitfan/commands/serviceInstall.go
@@ -1,4 +1,4 @@
-package cli
+package commands
 
 import (
 	"fmt"
