+++ cmd/bitfan/commands/serviceStart.go
@@ -1,4 +1,4 @@
-package cli
+package commands
 
 import (
 	"log"
