+++ cmd/bitfan/commands/serviceStop.go
@@ -1,4 +1,4 @@
-package cli
+package commands
 
 import (
 	"log"
