+++ cmd/bitfan/commands/stop.go
@@ -1,4 +1,4 @@
-package cli
+package commands
 
 import (
 	"fmt"
